% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{checkStrata}
\alias{checkStrata}
\title{A function to check if the composition and fallback strata are
compatible. Each composition stratum must have one and only one fallback stratum}
\usage{
checkStrata(stratAssign_comp, stratAssign_fallback)
}
\arguments{
\item{stratAssign_comp}{the strata assignment tibble for composition estimates}

\item{stratAssign_fallback}{the strata assignment tibble for fallback/reascension estimates}
}
\description{
A function to check if the composition and fallback strata are
compatible. Each composition stratum must have one and only one fallback stratum
}
