% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night_fall_reascend_wc_binom.R
\name{expand_wc_binom_night}
\alias{expand_wc_binom_night}
\title{estimate and bootstrap window count as a binomial and expanding for nighttime passage}
\usage{
expand_wc_binom_night(
  nightPassage_rates,
  wc,
  wc_prop,
  stratAssign_comp,
  stratAssign_night,
  boots = 2000
)
}
\arguments{
\item{nightPassage_rates}{one of the outputs of \code{nightFall}}

\item{wc}{tibble with two columns sWeek and wc. sWeek is statistical week and wc is the count of fish (NOT expanded for wc_prop)}

\item{wc_prop}{the proportion of the time fish are counted (e.g. 5/6)}

\item{stratAssign_comp}{tibble with sWeek, stratum showing what stratum each sWeek corresponds to
for composition estimation using the trap data}

\item{stratAssign_night}{tibble with sWeek, stratum showing what stratum each sWeek corresponds to
for nighttime passage}
}
\description{
estimate and bootstrap window count as a binomial and expanding for nighttime passage
}
