% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNC_expand_one_strat.R
\name{HNC_expand_one_strat}
\alias{HNC_expand_one_strat}
\title{one stratum, one dataset composition estimation, accounting}
\usage{
HNC_expand_one_strat(
  trap,
  H_vars,
  HNC_vars,
  W_vars,
  wc_expanded,
  pbt_var = NULL,
  tagRates = NULL
)
}
\arguments{
\item{trap}{data for ONE stratum}
}
\description{
one stratum, one dataset composition estimation, accounting
}
