% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNC_expand_one_strat.R
\name{HNC_expand_one_strat_MLE}
\alias{HNC_expand_one_strat_MLE}
\title{one stratum, one dataset composition estimation, MLE
treating phystag and phys-untag separately
stepwise to deal with consistency and missing data issues}
\usage{
HNC_expand_one_strat_MLE(
  trap,
  H_vars,
  HNC_vars,
  W_vars,
  wc_expanded,
  pbt_var = NULL,
  tagRates = NULL
)
}
\arguments{
\item{trap}{data for ONE stratum}
}
\description{
one stratum, one dataset composition estimation, MLE
treating phystag and phys-untag separately
stepwise to deal with consistency and missing data issues
}
