% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_HNC_expand.R
\name{HNC_expand_unkGSI}
\alias{HNC_expand_unkGSI}
\title{Treats GSI as unknown}
\usage{
HNC_expand_unkGSI(
  trap,
  stratAssign_comp,
  boots = 2000,
  pbt_var,
  timestep_var,
  physTag_var,
  adclip_var,
  sampID,
  tagRates,
  H_vars,
  HNC_vars,
  W_vars,
  wc_binom,
  GSI_draws,
  n_point = 100,
  GSI_var,
  method = c("Account", "MLE")
)
}
\arguments{
\item{GSI_draws}{a tibble wiht the first column individuals names, the rest GSI
values fro draws from the posterior (each column a draw). Used in order.}

\item{n_point}{number of iterations in the bootstrap loop (only resampling GSI values)
for calculating point estimates}

\item{GSI_var}{the column name in \code{trap} that has the GSI values.}
}
\description{
Treats GSI as unknown
}
