% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{splitByPIT}
\alias{splitByPIT}
\title{a function to split PBT release groups estimated by var1
according to PIT tag detections. Used for WA "release groups"
that are split between groups released below and above LGD.}
\usage{
splitByPIT(est_comp, PITdetectionSummary)
}
\arguments{
\item{est_comp}{The output of either \code{ascension_composition} or
\code{HNC_expand} where PBT release group was estimated as var1 for H and HNC.}

\item{PITdetectionSummary}{A tibble with columns releaseGroupPBT, stockGroup,
releaseGroupPIT (optional), detectPIT, and tagRatePIT. Each line defines a
PIT tag release group. releaseGroupPBT is the var1 (within H and HNC) that the PIT
release group belongs to. stockGroup defines the stockGroup it belongs to (typically "upper" or "lower")
detectPIT is the number of PIT tags detected from this group (actual number NOT expanded number)
and tagRatePIT is the PIT tagging rate for this group.}
}
\description{
a function to split PBT release groups estimated by var1
according to PIT tag detections. Used for WA "release groups"
that are split between groups released below and above LGD.
}
