% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_HNC_expand.R
\name{HNC_expand}
\alias{HNC_expand}
\title{treats GSI as known}
\usage{
HNC_expand(
  trap,
  stratAssign_comp,
  boots = 2000,
  pbt_var,
  timestep_var,
  physTag_var,
  adclip_var,
  tagRates,
  H_vars,
  HNC_vars,
  W_vars,
  wc_binom,
  method = c("Account", "MLE")
)
}
\description{
treats GSI as known
}
