% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night_fall_reascend_wc_binom.R
\name{nightFall}
\alias{nightFall}
\title{estimate fallback, reascension, and nighttime passage}
\usage{
nightFall(
  full_reascend,
  full_night,
  stratAssign_fallback,
  stratAssign_night,
  boots = 2000,
  full_spillway = NULL
)
}
\arguments{
\item{full_reascend}{tibble with sWeek, stockGroup, numReascend, totalPass. sWeek is statistical week
stockGroup is the stockGroup (fallback/reascension rates different between stock groups), numReascend is either the
number of PITs ascending that were reascensions in that sWeek or the number of PITs ascending that later
reascended (if using spillway data it MUST be the second)}

\item{full_night}{tibble with sWeek, nightPass, totalPass}

\item{stratAssign_fallback}{tibble with sWeek, stockGroup, stratum showing what stratum each sWeek
corresponds to for each stockGroup}

\item{stratAssign_night}{tibble with sWeek, stratum showing what stratum each sWeek corresponds to
for nighttime passage}

\item{boots}{The number of bootstap iterations to run}

\item{full_spillway}{tibble with "sWeek", "stockGroup", "laterAscend", "totalFall" columns. totalFall
gives the number of PIT ascensions that were detected (with the spillway array) falling back and laterAscend
gives the number of these that later reascended. If a strata has a summed value of 0 for totalFall,
it will be assumed that all fallbacks result in reascension in that stratum. If a value of NULL is input
for this, it will be assume that all fallbakcs result in reascension in all strata.}
}
\description{
estimate fallback, reascension, and nighttime passage
}
