% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night_fall_reascend.R
\name{nightFall}
\alias{nightFall}
\title{estimate fallback, reascension, and nighttime passage}
\usage{
nightFall(
  wc,
  full_reascend,
  full_night,
  stratAssign_fallback,
  stratAssign_night,
  full_spillway = NULL
)
}
\arguments{
\item{stratAssign_fallback}{tibble with sWeek, stockGroup, stratum showing what stratum each sWeek
corresponds to for each stockGroup}

\item{stratAssign_night}{tibble with sWeek, stratum showing what stratum each sWeek corresponds to
for nighttime passage}
}
\description{
estimate fallback, reascension, and nighttime passage
}
